Refactor IntegerParser into IntegerSetting
